Require JWTReader client to specify crypto alg
and wrap it up with the optional secret - because the secret is empty
if and only if the algorithm is NoAlgorithm.

Also check that the algorithm matches the expected one if there is a
signature.

No new tests needed because, as there are only two algorithms supported
at present, the existing test cases entirely cover this.